+++ CHANGELOG.md
@@ -52,6 +52,7 @@ of `major.minor.patch` with
 - Variable `depth` initialized to zero before layer loop in `fwriteoutput.c` to correct `soilc_1m` output.
 - Length of buffer for hostname corrected in `gethost.c`.
 - Output of unprintable characters corrected in `fputprintable.c`.
+- Missing path to grid file added in `openmetafile.c`.
 
 
 ## [6.0.7] - 2026-04-17